+++ mitmproxy/certs.py
@@ -49,6 +49,9 @@ class Cert(serializable.Serializable):
     def __eq__(self, other):
         return self.fingerprint() == other.fingerprint()
 
+    def __repr__(self):
+        return f"<Cert(cn={self.cn!r}, altnames={self.altnames!r})>"
+
     @classmethod
     def from_state(cls, state):
         return cls.from_pem(state)